Update main.c
Fix assert error (Line 399)
Description: The issue was that the assert will test if the first element in adjacencyList is empty, which was incorrect.

Remove redundant code in stage 1.
Line 611:  if block was removed as the loop will be entered  if and only if the condition was meet in davance.

3. Remove function resetSandbox (Line 627)
Description: remove this function as it is no longer required.